switched to iMultiFabs for DOF objects

+++ Source/FieldSolver/ImplicitSolvers/WarpXSolverVec.cpp
@@ -158,7 +158,7 @@ void WarpXSolverVec::copyFrom ( const amrex::Real* const a_arr)
                     ParallelFor( bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                     {
                         for (int v = 0; v < ncomp; v++) {
-                            int dof = (int) dof_arr(i,j,k,2*v); // local
+                            int dof = dof_arr(i,j,k,2*v); // local
                             data_arr(i,j,k,v) = a_arr[dof];
                         }
                     });
@@ -174,7 +174,7 @@ void WarpXSolverVec::copyFrom ( const amrex::Real* const a_arr)
                 ParallelFor( bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                 {
                     for (int v = 0; v < ncomp; v++) {
-                        int dof = (int) dof_arr(i,j,k,2*v); // local
+                        int dof = dof_arr(i,j,k,2*v); // local
                         data_arr(i,j,k,v) = a_arr[dof];
                     }
                 });
@@ -203,7 +203,7 @@ void WarpXSolverVec::copyTo ( amrex::Real* const a_arr) const
                     ParallelFor( bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                     {
                         for (int v = 0; v < ncomp; v++) {
-                            int dof = (int) dof_arr(i,j,k,2*v); // local
+                            int dof = dof_arr(i,j,k,2*v); // local
                             a_arr[dof] = data_arr(i,j,k,v);
                         }
                     });
@@ -219,7 +219,7 @@ void WarpXSolverVec::copyTo ( amrex::Real* const a_arr) const
                 ParallelFor( bx, [=] AMREX_GPU_DEVICE (int i, int j, int k)
                 {
                     for (int v = 0; v < ncomp; v++) {
-                        int dof = (int) dof_arr(i,j,k,2*v); // local
+                        int dof = dof_arr(i,j,k,2*v); // local
                         a_arr[dof] = data_arr(i,j,k,v);
                     }
                 });
